refactor: components, utils 폴더 생성
