+++ src/components/CleanRateBox.jsx
