+++ src/pages/MyPage/SetType.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
-import SetHeader from "./SetHeader";
-import { Container } from "../../theme/SharedContainer";
+import SharedHeader from "../../components/SharedHeader";
+import { Container } from "../../components/SharedContainer";
 import smokerImg from "../../assets/smoker.png";
 import nonSmokerImg from "../../assets/nonSmoker.png";
 import detailBackgroundImage from '../../assets/mypage_detail_background.png';
@@ -32,7 +32,7 @@ const SetType = () => {
 
   return (
     <SelectContainer>
-      <SetHeader headerText="사용자 유형 변경"></SetHeader>
+      <SharedHeader headerText="사용자 유형 변경"></SharedHeader>
       <SmokeSelect>
         <SmokerBox
           className={activeBox === "smoker" ? "active" : ""}
